Add -f switch, dont remove containers

+++ run.sh
@@ -2,13 +2,30 @@
 
 cd $(dirname $0)
 
-SOURCE=$(pwd)/android
-CCACHE=$(pwd)/ccache
+SOURCE=$(readlink -f $(pwd)/android)
+CCACHE=$(readlink -f $(pwd)/ccache)
 CONTAINER_HOME=/home/cmbuild
 CONTAINER=cyanogenmod
 REPOSITORY=stucki/cyanogenmod
 TAG=cm-13.0
 FORCE_BUILD=0
+
+
+while [[ $# > 0 ]]
+do
+key="$1"
+
+case $key in
+    -r|--rebuild)
+    FORCE_BUILD=1
+    ;;
+    *)
+    shift # past argument or value
+    ;;
+esac
+shift
+
+done
 
 # Create shared folders
 mkdir -p $SOURCE
@@ -26,12 +43,17 @@ elif [[ $FORCE_BUILD = 1 ]] || ! echo "$IMAGE_EXISTS" | grep -q "$TAG"; then
 
 	echo "Building Docker image $REPOSITORY:$TAG..."
 	docker build -t $REPOSITORY:$TAG .
+	OK=$?
 
 	# After successful build, delete existing containers
 	IS_EXISTING=$(docker inspect -f '{{.Id}}' $CONTAINER 2>/dev/null)
-	if [[ -n "$IS_EXISTING" ]]; then
+	if [[ $OK -eq 0 ]] && [[ -n "$IS_EXISTING" ]]; then
 		docker rm $CONTAINER
 	fi
+fi
+
+if [[ $OK -ne 0 ]]; then
+	exit 1;
 fi
 
 # With the given name $CONTAINER, reconnect to running container, start
